+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 - Show the type of change in the overlapping code area's title.
 - When new and deleted files are empty, show `<contents empty>` in code area.
 - Use ▓ characters to demarcate diff sections when showing in code area.
-- Show authored and committed date information.
+- Show authored and committed date information (along with natural time).
 
 ### Changed
 - Automatically scroll to the top of the code area after selecting a new file or revision.